+++ lato/testing.py
@@ -1,5 +1,5 @@
 import contextlib
-from typing import Iterator
+from collections.abc import Iterator
 
 from lato import Application